Fix ValueError: Attempted relative import in non-package
